requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-allauth==0.61.0
django-allauth[google]==0.61.0
django-colorfield==0.11.0
django-environ==0.11.2
django-jalali-date==1.1.2
django-rosetta==0.10.0
django-smtp-ssl==1.0
django-widget-tweaks==1.5.0
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.3.0
polib==1.2.0
pycparser==2.22
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
